refactor: useNotifications 사용
