+++ src/components/domain/AlarmModal/index.jsx
@@ -1,8 +1,11 @@
-import { getNotifications, updateSeenNotification } from '@/api/alarm';
+import { queryClient } from '@/App';
 
-import { Heading, Icon, Text } from '@/components/base';
+import { updateSeenNotification } from '@/api/notification';
+
+import { Heading, Icon, Spinner, Text } from '@/components/base';
 import SettingModal from '@/components/domain/SettingModal';
 
+import { useGetNotifications } from '@/hooks/queries/notifications';
 import useInput from '@/hooks/useInput';
 
 import { convertDate } from '@/utils/date';
@@ -25,19 +28,11 @@ import {
 
 function AlarmModal({ visible, onClose }) {
   const navigate = useNavigate();
+  const { data: alarms, isLoading } = useGetNotifications();
 
-  const [alarms, setAlarms] = useState([]);
   const [filteredAlarms, setFilteredAlarms] = useState(alarms);
   const clickedIndex = useInput(0);
   const [settingModalVisible, setSettingModalVisible] = useState(false);
-
-  useEffect(() => {
-    if (!visible) return;
-
-    (async () => {
-      setAlarms(await getNotifications());
-    })();
-  }, [visible]);
 
   const handleAlarmClick = (postId) => {
     navigate(`/TIL/${postId}`);
@@ -48,7 +43,8 @@ function AlarmModal({ visible, onClose }) {
     if (alarms.every(({ seen }) => seen)) return;
 
     await updateSeenNotification();
-    setAlarms(await getNotifications());
+
+    queryClient.invalidateQueries(['notifications']);
   };
 
   useEffect(() => {
@@ -85,29 +81,33 @@ function AlarmModal({ visible, onClose }) {
         </StyledHeaderItem>
         <Icon size='medium' isPointer={true} onClick={() => onClose && onClose()} />
       </StyledHeaderContainer>
-      <StyledAlarmContainer>
-        {alarms.length === 0 || (filteredAlarms.length === 0 && clickedIndex.value !== 2) ? (
-          <StyledNoAlarm>🥲 알림이 없어요...</StyledNoAlarm>
-        ) : filteredAlarms.length === 0 ? (
-          <StyledNoAlarm>🥲 읽지 않은 알림이 없어요...</StyledNoAlarm>
-        ) : (
-          filteredAlarms.map(({ _id, author, post, like, comment }) => (
-            <StyledAlarm key={_id} onClick={() => handleAlarmClick(post)}>
-              <StyledAvatar src={author.image} size='medium' />
-              <StyledContentContainer>
-                <StyledContent size='medium'>
-                  {like
-                    ? `${author.fullName}님이 ${like.post.title.title}을(를) 좋아합니다.`
-                    : `${author.fullName}님이 ${comment.post.title.title}에 남긴 댓글: ${comment.comment}`}
-                </StyledContent>
-                <Text size='small' color={COLOR.CREATEDAT}>
-                  {convertDate(new Date(like ? like.post.updatedAt : comment.post.updatedAt))}
-                </Text>
-              </StyledContentContainer>
-            </StyledAlarm>
-          ))
-        )}
-      </StyledAlarmContainer>
+      {isLoading ? (
+        <Spinner />
+      ) : (
+        <StyledAlarmContainer>
+          {alarms.length === 0 || (filteredAlarms.length === 0 && clickedIndex.value !== 2) ? (
+            <StyledNoAlarm>🥲 알림이 없어요...</StyledNoAlarm>
+          ) : filteredAlarms.length === 0 ? (
+            <StyledNoAlarm>🥲 읽지 않은 알림이 없어요...</StyledNoAlarm>
+          ) : (
+            filteredAlarms.map(({ _id, author, post, like, comment }) => (
+              <StyledAlarm key={_id} onClick={() => handleAlarmClick(post)}>
+                <StyledAvatar src={author.image} size='medium' />
+                <StyledContentContainer>
+                  <StyledContent size='medium'>
+                    {like
+                      ? `${author.fullName}님이 ${like.post.title.title}을(를) 좋아합니다.`
+                      : `${author.fullName}님이 ${comment.post.title.title}에 남긴 댓글: ${comment.comment}`}
+                  </StyledContent>
+                  <Text size='small' color={COLOR.CREATEDAT}>
+                    {convertDate(new Date(like ? like.post.updatedAt : comment.post.updatedAt))}
+                  </Text>
+                </StyledContentContainer>
+              </StyledAlarm>
+            ))
+          )}
+        </StyledAlarmContainer>
+      )}
     </StyledAlarmModal>
   );
 }
